refactor(baselines): hoist readPriorBaselineRows into shared helper
Extract the prior-baseline reader from update-maintainability-baseline
and update-crap-baseline into a single readPriorBaselineRows() in
diff-scope-cli.js, handling both envelope and legacy flat-map shapes.
Brings each CLI back under its pre-#1974 maintainability baseline.
